Rename distributionUtils into distribution_utils
