+++ bin/distribute_elasticc.py
@@ -31,7 +31,7 @@ from fink_science import __version__ as fsvsn
 
 from fink_broker.parser import getargs
 from fink_broker.sparkUtils import init_sparksession, connect_to_raw_database
-from fink_broker.distributionUtils import get_kafka_df
+from fink_broker.distribution_utils import get_kafka_df
 from fink_broker.loggingUtils import get_fink_logger, inspect_application
 from fink_broker.partitioning import convert_to_millitime
